LayoutTests/fast/viewport/viewport-79.html

<head>
    <title>Semi-colon is recognized as a property value.</title>
    <meta name="viewport" content="width=100 a=; initial-scale=1">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
