safari-600.7-branch/LayoutTests/fast/viewport/viewport-79-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "a" not recognized and ignored.
ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
